requirements.txt
selenium~=4.19.0
pyautogui
Pillow
instructor
openai
python-dotenv
anthropic